event-timing/shadow-dom-null-target.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=/resources/testdriver-actions.js></script>
<script src=resources/event-timing-test-utils.js></script>
<body>
<div id='host'>
</div>
<script>
promise_test(async t => {
  assert_implements(window.PerformanceEventTiming, 'Event Timing is not supported.');

  const durationThreshold = 16;

  const host = document.getElementById('host');
  const shadowRoot = host.attachShadow({mode: "closed"});
  const innerSpan = document.createElement("span");
  innerSpan.textContent = "Span";
  innerSpan.addEventListener("mouseover", function(e) {
    // To make this is a slow event
    mainThreadBusy(durationThreshold + 4);
  });
  shadowRoot.append(innerSpan);

  const spanRect = innerSpan.getBoundingClientRect();

  // Create Observer before sending mousemove action.
  const observerPromise = new Promise(resolve => {
    new PerformanceObserver(t.step_func(entryList => {
      let mouseOverEntries = entryList.getEntriesByName("mouseover");
      assert_equals(mouseOverEntries.length, 1);
      assert_equals(mouseOverEntries[0].target, null);
      resolve();
    })).observe({type: 'event', durationThreshold: durationThreshold});
  });

  // Hover over the span inside the shadow dom.
  const actions = new test_driver.Actions();
  await actions.pointerMove(
    Math.ceil(spanRect.x + spanRect.width / 2),
    Math.ceil(spanRect.y + spanRect.height / 2)
  ).send();

  await observerPromise;
}, "Event Timing: Move pointer within shadow DOM should create event-timing entry with null target.");
</script>
</body>
</html>
