third_party/blink/web_tests/platform/mac/external/wpt/event-timing/shadow-dom-null-target-expected.txt

This is a testharness.js-based test.
[FAIL] Event Timing: Move pointer within shadow DOM should create event-timing entry with null target.
  assert_equals: expected null but got Element node <div id="host">\n</div>
Harness: the test ran to completion.
